graphics bugs-
1. multriple tries to get it running
2. pixelated

+++ app/src/main/cpp/GameBoyCore.h
@@ -5,6 +5,6 @@
 
 FORGE_API void       startEmulator ();
 FORGE_API void       stopEmulator  ();
-FORGE_API void       getFrame      (u16* FRAME_BUFFER);
+FORGE_API void       getFrame      (u32* FRAME_BUFFER);
 FORGE_API void       getAudio      (u8* AUDIO_BUFFER);
 FORGE_API void       setButton     (u8 BUTTON, bool PRESSED);
